requirements.txt
google-adk==1.14.0
python-dotenv
langchain-community==0.3.27
wikipedia==1.4.0
google-cloud-logging
